Degrains: compute DegrainWeight numerator in double to avoid int64 overflow
(thSAD - blockSAD) * (thSAD + blockSAD) * 256 was evaluated in int64
before the double divide and overflows once thSAD exceeds ~1.9e8. The
create-time guard only rejects thSAD >= INT_MAX (~2.1e9), so the range in
between (reachable with high bit depth + large blocks scaling thsad by
~6.5e4) passed validation but produced signed-overflow garbage weights.
Promote the numerator (and denominator) operands to double.

+++ src/Degrains.h
@@ -188,7 +188,10 @@ static inline int DegrainWeight(int64_t thSAD, int64_t blockSAD) {
     if (blockSAD >= thSAD)
         return 0;
 
-    return int((thSAD - blockSAD) * (thSAD + blockSAD) * 256 / (double)(thSAD * thSAD + blockSAD * blockSAD));
+    // Compute in double: (thSAD - blockSAD) * (thSAD + blockSAD) * 256 overflows int64 once
+    // thSAD exceeds ~1.9e8 (reachable with high bit depth + large blocks + a large thsad,
+    // all below the INT_MAX create-time guard), producing garbage weights.
+    return int((double)(thSAD - blockSAD) * (double)(thSAD + blockSAD) * 256.0 / ((double)thSAD * thSAD + (double)blockSAD * blockSAD));
 }
 
 template<typename PixelType>
